+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -10,21 +10,9 @@
         android:name="com.example.chotuvemobileapp.ui.home.HomeFragment"
         android:label="@string/menu_home"
         tools:layout="@layout/fragment_home">
-
         <action
-            android:id="@+id/action_HomeFragment_to_HomeSecondFragment"
-            app:destination="@id/nav_home_second" />
-    </fragment>
-    <fragment
-        android:id="@+id/nav_home_second"
-        android:label="@string/home_second">
-        <action
-            android:id="@+id/action_HomeSecondFragment_to_HomeFragment"
-            app:destination="@id/nav_home" />
-
-        <argument
-            android:name="myArg"
-            app:argType="string" />
+            android:id="@+id/action_nav_home_to_addVideoFragment"
+            app:destination="@id/addVideoFragment" />
     </fragment>
 
     <fragment
@@ -38,4 +26,13 @@
         android:name="com.example.chotuvemobileapp.ui.slideshow.SlideshowFragment"
         android:label="@string/menu_slideshow"
         tools:layout="@layout/fragment_slideshow" />
+    <fragment
+        android:id="@+id/addVideoFragment"
+        android:name="com.example.chotuvemobileapp.ui.home.AddVideoFragment"
+        android:label="fragment_add_video"
+        tools:layout="@layout/fragment_add_video" >
+        <action
+            android:id="@+id/action_addVideoFragment_to_nav_home"
+            app:destination="@id/nav_home" />
+    </fragment>
 </navigation>
